refactor Popup component
